ags: try to fix better_multimonitor.js
Now that I actually have a second monitor I tried to fix this but I ran
into some really weird async shit with the monitor not appearing in the
hyprland service even 5 seconds after the monitor has been added. I'm
sure this is something dumb but I don't have the time to fix it rn.